added a test showing the case
